change name of getDapps v3 to chaindapps and refactor v3 controller
